# Antigravity/docs - Copy/design/api/openapi.yaml
openapi: 3.0.3
info:
  title: Honda DMS API
  description: |
    Honda Dealer Management System - RESTful API Specification
    
    ## Overview
    Complete API for managing Honda dealership operations including:
    - Customer Relationship Management (CRM)
    - Sales & Quotations
    - Service & Repair Orders
    - Parts Inventory
    - Insurance & Accounting
    
    ## Authentication
    All endpoints require JWT Bearer token authentication.
    
    ## Rate Limiting
    - 1000 requests per hour per user
    - 100 requests per minute per user
    
    ## Versioning
    Current version: v1.0
    
  version: 1.0.0
  contact:
    name: Honda DMS API Support
    email: api@hondadms.com
  license:
    name: Proprietary
servers:
  - url: https://api.hondadms.com/v1
    description: Production server
  - url: https://staging-api.hondadms.com/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
...
security:
  - bearerAuth: []
paths:
...
  /sales/quotations:
    get:
      tags: [Sales - Quotations]
      summary: List quotations
      operationId: listQuotations
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Quotations retrieved successfully
    
    post:
      tags: [Sales - Quotations]
      summary: Create quotation
      operationId: createQuotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuotationRequest'
      responses:
        '201':
          description: Quotation created successfully

  # ==================== SERVICE - REPAIR ORDERS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    CreateQuotationRequest:
      type: object
      required: [customer_name, customer_phone, model, version, color, base_price, total_price]
      properties:
        customer_id:
          type: string
        customer_name:
          type: string
        customer_phone:
          type: string
        model:
          type: string
        version:
          type: string
        color:
          type: string
        base_price:
          type: number
        accessories:
          type: array
          items:
            type: string
        total_price:
          type: number
...